refactor(presentation): split dto wire-types from From-mapping impls (#79)
